+++ src/components/index.ts
@@ -21,3 +21,4 @@ export * from './Model';
 export * from './Plugins';
 export * from './Switch';
 export * from './Toast';
+export * from './Popup';